+++ src/curtain-accessory.ts
@@ -10,20 +10,182 @@ import {
   HAP,
   Logging,
   Service,
-  CharacteristicEventTypes,
+  CharacteristicEventTypes, Characteristic,
 } from "homebridge";
+
+type OnPositionReceived = (pos: number) => void
+
+class Device {
+  private readonly bleMac: string;
+  private readonly scanDuration: number;
+  private readonly slowScanInterval: number;
+  private readonly fastScanInterval = 500;
+  private readonly fastScanDuration = 10000;
+
+  private readonly name: string;
+  private readonly log: Logging;
+  private readonly callback: OnPositionReceived;
+
+  private openCloseThreshold = 5;
+
+  private fastScanEnabled = false;
+  private autoDisableFastScanTimeoutId: NodeJS.Timeout | null = null;
+
+  private scanIntervalId: NodeJS.Timeout | null = null;
+
+  private previousPosition = -1;
+  private position = -1;
+
+  constructor(bleMac: string, name: string, log: Logging, scanDuration: number, scanInterval: number, openCloseThreshold: number,
+    callback: OnPositionReceived) {
+    this.name = name;
+    this.bleMac = bleMac;
+    this.log = log;
+
+    this.scanDuration = scanDuration;
+    this.slowScanInterval = scanInterval;
+    this.callback = callback;
+
+    this.openCloseThreshold = openCloseThreshold;
+
+    this.startFastScan();
+  }
+
+  /**
+   * Ask the device to start moving to the given position (which must be a device value and not an HomeKit value).
+   * Returns a Promise that's resolved as soon as the command was sent to the device.
+   */
+  public runToPosition(pos: number): Promise<void> {
+    const SwitchBot = require("node-switchbot");
+    const switchbot = new SwitchBot();
+
+    return switchbot
+      .discover({duration: this.scanDuration, model: "c", quick: false})
+      .then((device_list: any) => {
+        let targetDevice: any = null;
+
+        for (let device of device_list) {
+          this.log.info(device.modelName, device.address);
+          if (device.address == this.bleMac) {
+            targetDevice = device;
+            break;
+          }
+        }
+
+        if (!targetDevice) {
+          return new Promise((resolve, reject) => {
+            reject(new Error("Curtain '" + this.name + "' (" + this.bleMac + "): device not found."));
+          });
+        }
+
+        this.startFastScan();
+
+        this.log.info("Curtain '%s' (%s) is moving to %d%%...", this.name, this.bleMac, pos);
+        return targetDevice.runToPos(pos);
+      });
+  }
+
+  /**
+   * Start a faster scan loop (using fastScanInterval). Used in calls to `runToPosition` to
+   * report quicker on device position change. Will disable on its own as soon if the curtain's
+   * device position does not change for more than `fastScanDuration`.
+   */
+  private startFastScan() {
+    if (this.fastScanEnabled) {
+      return;
+    }
+    this.fastScanEnabled = true;
+    this.startScanLoop();
+  }
+
+  private stopFastScan() {
+    if (!this.fastScanEnabled) {
+      return;
+    }
+    this.fastScanEnabled = false;
+    this.startScanLoop();
+  }
+
+  private get scanInterval(): number {
+    return this.fastScanEnabled ? this.fastScanInterval : this.slowScanInterval;
+  }
+
+  private startScanLoop() {
+    if (this.scanIntervalId !== null) {
+      clearInterval(this.scanIntervalId);
+    }
+
+    this.log.info("Curtain '%s': starting scan loop with interval %d", this.name, this.scanInterval);
+
+    this.scanIntervalId = setInterval(() => {
+      this.scan().catch((err) => {
+        this.log.error("error while scanning for Curtain '%s': %s", this.name, err);
+      });
+    }, this.scanInterval);
+  }
+
+  private scan(): Promise<void> {
+    return new Promise((resolve, reject) => {
+      const SwitchBot = require("node-switchbot");
+      const switchbot = new SwitchBot();
+      switchbot.onadvertisement = (ad: any) => {
+        this.applyPosition(ad);
+      };
+      switchbot.startScan({id: this.bleMac})
+        .then(() => {
+          return switchbot.wait(this.scanDuration);
+        })
+        .then(() => {
+          resolve();
+          switchbot.stopScan();
+        })
+        .catch((err: any) => {
+          reject(err);
+        });
+    });
+  }
+
+  private applyPosition(ad: any) {
+    let pos = ad.serviceData.position;
+
+    if (pos + this.openCloseThreshold >= 100) {
+      pos = 100;
+    } else if (pos - this.openCloseThreshold <= 0) {
+      pos = 0;
+    }
+
+    if (pos === this.previousPosition) {
+      return;
+    }
+
+    this.previousPosition = this.position;
+    this.position = pos;
+
+    if (this.autoDisableFastScanTimeoutId !== null) {
+      clearTimeout(this.autoDisableFastScanTimeoutId);
+    }
+
+    this.autoDisableFastScanTimeoutId = setTimeout(() => {
+      this.stopFastScan();
+    }, this.fastScanDuration);
+
+    this.callback(this.position);
+  }
+}
 
 export class Curtain implements AccessoryPlugin {
   private readonly log: Logging;
-  private readonly bleMac: string;
-  private readonly scanDuration: number;
   private readonly reverseDir: boolean;
   private readonly moveTime: number;
+
+  private didReceiveInitialPosition = false;
 
   private currentPosition = 0;
   private targetPosition = 0;
   private positionState = 0;
   private moveTimer!: NodeJS.Timeout;
+
+  private device: Device;
 
   // This property must be existent!!
   name: string;
@@ -31,26 +193,62 @@ export class Curtain implements AccessoryPlugin {
   private readonly curtainService: Service;
   private readonly informationService: Service;
 
-  constructor(hap: HAP, log: Logging, name: string, bleMac: string, scanDuration: number, reverseDir: boolean, moveTime: number) {
+  private currentPositionCharacteristic: Characteristic;
+  private targetPositionCharacteristic: Characteristic;
+  private positionStateCharacteristic: Characteristic;
+
+  constructor(hap: HAP, log: Logging, name: string, bleMac: string,
+    scanDuration: number, reverseDir: boolean, moveTime: number, scanInterval: number, openCloseThreshold: number) {
     this.log = log;
     this.name = name;
-    this.bleMac = bleMac;
-    this.scanDuration = scanDuration;
     this.reverseDir = reverseDir;
     this.moveTime = moveTime;
 
+    this.device = new Device(bleMac, name, log, scanDuration, scanInterval, openCloseThreshold, (pos: number) => {
+      this.currentPosition = this.convertFromHomeKitPosition(pos);
+      this.currentPositionCharacteristic.updateValue(this.currentPosition);
+
+      // Set target to the same value as current on the first received position, or if the curtain is in
+      // STOPPED state (eg. it's still or not moving from a change triggered by HomeKit). Otherwise, with
+      // target = X and current != X, HomeKit will assume curtain is moving and show the progress indicator.
+      //
+      // Note that on an HomeKit-triggered change, `positionState` is set to STOPPED after `moveTime` has
+      // passed, rather than when the curtain actually finished moving. When `moveTime` is lower than the
+      // curtain's actual travel time, the device will appear in HomeKit first as moving (with a progress
+      // indicator) then after `moveTime` has passed, as not moving (without progress indicator), but with
+      // rapidly changing percentages.
+      if (!this.didReceiveInitialPosition || this.positionState === hap.Characteristic.PositionState.STOPPED) {
+        this.didReceiveInitialPosition = true;
+        this.targetPosition = this.currentPosition;
+        this.targetPositionCharacteristic.updateValue(this.targetPosition);
+      }
+    });
+
     this.positionState = hap.Characteristic.PositionState.STOPPED;
 
+    this.informationService = new hap.Service.AccessoryInformation()
+      .setCharacteristic(hap.Characteristic.Manufacturer, "SwitchBot")
+      .setCharacteristic(hap.Characteristic.Model, "SWITCHBOT-CURTAIN-W0701600")
+      .setCharacteristic(hap.Characteristic.SerialNumber, bleMac);
+
     this.curtainService = new hap.Service.WindowCovering(name);
-    this.curtainService
-      .getCharacteristic(hap.Characteristic.CurrentPosition)
+
+    this.currentPositionCharacteristic = this.curtainService
+      .getCharacteristic(hap.Characteristic.CurrentPosition);
+
+    this.targetPositionCharacteristic = this.curtainService
+      .getCharacteristic(hap.Characteristic.TargetPosition);
+
+    this.positionStateCharacteristic = this.curtainService
+      .getCharacteristic(hap.Characteristic.PositionState);
+
+    this.currentPositionCharacteristic
       .on(CharacteristicEventTypes.GET, (callback: CharacteristicGetCallback) => {
         log.info("Current position of Curtain was returned: " + this.currentPosition + "%");
         callback(undefined, this.currentPosition);
       });
 
-    this.curtainService
-      .getCharacteristic(hap.Characteristic.TargetPosition)
+    this.targetPositionCharacteristic
       .on(CharacteristicEventTypes.GET, (callback: CharacteristicGetCallback) => {
         log.info("Target position of Curtain was returned: " + this.targetPosition + "%");
         callback(undefined, this.targetPosition);
@@ -68,64 +266,21 @@ export class Curtain implements AccessoryPlugin {
         }
 
         if (this.positionState === hap.Characteristic.PositionState.STOPPED) {
-          this.curtainService?.getCharacteristic(hap.Characteristic.TargetPosition).updateValue(this.targetPosition);
-          this.curtainService?.getCharacteristic(hap.Characteristic.CurrentPosition).updateValue(this.currentPosition);
-          this.curtainService?.getCharacteristic(hap.Characteristic.PositionState).updateValue(this.positionState);
+          this.targetPositionCharacteristic.updateValue(this.targetPosition);
+          this.currentPositionCharacteristic.updateValue(this.currentPosition);
+          this.positionStateCharacteristic.updateValue(this.positionState);
           callback();
         } else {
-          const SwitchBot = require("node-switchbot");
-          const switchbot = new SwitchBot();
-          switchbot
-            .discover({ duration: this.scanDuration, model: "c", quick: false })
-            .then((device_list: any) => {
-              log.info("Scan done.");
-              let targetDevice: any = null;
-              for (let device of device_list) {
-                log.info(device.modelName, device.address);
-                if (device.address == this.bleMac) {
-                  targetDevice = device;
-                  break;
-                }
-              }
-              if (!targetDevice) {
-                log.info(name + " (" + bleMac + ") was not found.");
-                return new Promise((resolve, reject) => {
-                  reject(new Error("No target device was found."));
-                });
-              } else {
-                log.info(targetDevice.modelName + " (" + targetDevice.address + ") was found.");
-                // Set event handers
-                targetDevice.onconnect = () => {
-                  // log.info('Connected.');
-                };
-                targetDevice.ondisconnect = () => {
-                  // log.info('Disconnected.');
-                };
-                log.info("Curtain is moving...");
-                /**opend - 0% in HomeKit, 100% in Curtain device.
-                 * closed - 100% in HomeKit, 0% in Curtain device.
-                 * To keep the status synchronized, convert the percentage of homekit to the percentage of Curtain.
-                 */
-                let covertToDevicePosition = 0;
-                if (!reverseDir) {
-                  covertToDevicePosition = 100 - this.targetPosition;
-                } else {
-                  log.info('Reverse the "opened" and "closed" directions!');
-                  covertToDevicePosition = this.targetPosition;
-                }
-                return targetDevice.runToPos(covertToDevicePosition);
-              }
-            })
+          this.device.runToPosition(this.convertFromHomeKitPosition(this.targetPosition))
             .then(() => {
               log.info("Done.");
               log.info("Target position of Curtain has been set to: " + this.targetPosition + "%");
               this.moveTimer = setTimeout(() => {
                 // log.info("setTimeout", this.positionState.toString(), this.currentPosition.toString(), this.targetPosition.toString());
-                this.currentPosition = this.targetPosition;
                 this.positionState = hap.Characteristic.PositionState.STOPPED;
                 // this.curtainService?.getCharacteristic(hap.Characteristic.TargetPosition).updateValue(this.targetPosition);
-                this.curtainService?.getCharacteristic(hap.Characteristic.CurrentPosition).updateValue(this.currentPosition);
-                this.curtainService?.getCharacteristic(hap.Characteristic.PositionState).updateValue(this.positionState);
+                this.currentPositionCharacteristic.updateValue(this.currentPosition);
+                this.positionStateCharacteristic.updateValue(this.positionState);
               }, this.moveTime);
               callback();
             })
@@ -134,9 +289,9 @@ export class Curtain implements AccessoryPlugin {
               this.moveTimer = setTimeout(() => {
                 this.targetPosition = this.currentPosition;
                 this.positionState = hap.Characteristic.PositionState.STOPPED;
-                this.curtainService?.getCharacteristic(hap.Characteristic.TargetPosition).updateValue(this.targetPosition);
+                this.targetPositionCharacteristic.updateValue(this.targetPosition);
                 // this.curtainService?.getCharacteristic(hap.Characteristic.CurrentPosition).updateValue(this.currentPosition);
-                this.curtainService?.getCharacteristic(hap.Characteristic.PositionState).updateValue(this.positionState);
+                this.positionStateCharacteristic.updateValue(this.positionState);
               }, 1000);
               log.info("Target position of Curtain failed to be set to: " + this.targetPosition + "%");
               callback();
@@ -144,19 +299,30 @@ export class Curtain implements AccessoryPlugin {
         }
       });
 
-    this.curtainService
-      .getCharacteristic(hap.Characteristic.PositionState)
+    this.positionStateCharacteristic
       .on(CharacteristicEventTypes.GET, (callback: CharacteristicGetCallback) => {
         log.info("The position state of Curtain was returned: " + this.positionState);
         callback(undefined, this.positionState);
       });
 
-    this.informationService = new hap.Service.AccessoryInformation()
-      .setCharacteristic(hap.Characteristic.Manufacturer, "SwitchBot")
-      .setCharacteristic(hap.Characteristic.Model, "SWITCHBOT-CURTAIN-W0701600")
-      .setCharacteristic(hap.Characteristic.SerialNumber, this.bleMac);
-
     log.info("Curtain '%s' created!", name);
+  }
+
+
+  /**
+   * Convert to/from device/HomeKit's position, since:
+   *
+   * - opened is 0% in HomeKit and 100% in Curtain device.
+   * - closed is 100% in HomeKit, 0% in Curtain device.
+   */
+  convertFromHomeKitPosition(n: number): number {
+    let covertToDevicePosition: number;
+    if (this.reverseDir) {
+      covertToDevicePosition = n;
+    } else {
+      covertToDevicePosition = 100 - n;
+    }
+    return covertToDevicePosition;
   }
 
   /*